moved models to extra file
